+++ src/views/partials/listing.blade.php
@@ -1,1 +1,1 @@
-<a href="{{ action($listingAction) }}" class="pull-right btn btn-default"><i class="fa fa-list-alt"></i>Back to Listing</a>
+<a href="{{ $listingAction }}" class="pull-right btn btn-default"><i class="fa fa-list-alt"></i>Back to Listing</a>